refactor: Create dedicated services directory
The `DataDeletionService` has been moved to a `services` directory,
independent of controllers. This was done to reinforce that services
are independent of controllers, and can be used outside of them.
